Fix displaying users when there are more than 250
getUsers is now a recursive function, which adds users to the table
based on how many pages the X-Total-Pages header returns.

The first if-sentence in the getUsers function "then" part exists
because whenever an user is added, the table should first be emptied and
then all the users are added again to avoid double-users.